backend/requirements.txt
fastapi
uvicorn[standard]
PyJWT
mysql-connector-python
python-dotenv
